2pt conical stage for focal-point-outside case
A couple of annoyances here:

1) the prev vector_scale stage is not usable for masking, as NaN values can propagate through
   => switch to actual masking

2) for the outside case, we must select the min root when the gradient is flipped
   => split into two templated stages (_min, _max)

  (I'm not convinced that we need to flip the gradient for RP at all; we can investigate later)

+++ src/jumper/SkJumper.h
@@ -108,11 +108,11 @@ struct SkJumper_GradientCtx {
 };
 
 struct SkJumper_2PtConicalCtx {
-    float fMask[SkJumper_kMaxStride];
-    float fCoeffA,
-          fInvCoeffA,
-          fR0,
-          fDR;
+    uint32_t fMask[SkJumper_kMaxStride];
+    float    fCoeffA,
+             fInvCoeffA,
+             fR0,
+             fDR;
 };
 
 #endif//SkJumper_DEFINED
